make a dedicated pe.rs file
