+++ crates/linker/Cargo.toml
@@ -20,9 +20,10 @@ bumpalo = { version = "3.8.0", features = ["collections"] }
 clap = { version = "3.1.15", default-features = false, features = ["std", "color", "suggestions"] }
 iced-x86 = { version = "1.15.0", default-features = false, features = ["std", "decoder", "op_code_info", "instr_info"] }
 memmap2 = "0.5.3"
-object = { version = "0.29.0", features = ["read", "write"] }
 mach_object = "0.1"
 serde = { version = "1.0.130", features = ["derive"] }
 bincode = "1.3.3"
 target-lexicon = "0.12.3"
 tempfile = "3.2.0"
+# object = { version = "0.29.0", features = ["read", "write"] }
+object = { path = "/home/folkertdev/rust/object", features = ["read", "write"] }
